imageless imageobjects
Adds the final boolean noImage and a new constructor for this. Also adds a new constructor to physics object to accommodate. Also added a couple more visual options to imageobject: boolean render and visualize. Render determines whether or not to render the object (it will still exist physically), visualize will draw a square around the perimeter of the object over the objects image (the default color is magenta).

The extra features should help with debugging and graphics. The purpose of making it possible to not have an image is to allow for trigger boxes to be created without having to upload an image. Also, it can be used by pairing a imageless physical object with a imageobject so that the imageobject's collider doesn't have to govern the physics.